+++ src/include/rccl_vars.h
@@ -27,6 +27,7 @@ THE SOFTWARE.
 
 RCCL_PARAM_DECLARE(EnableHipGraph);  // Opt-in environment variable for enabling hipGraph
 
+#define RCCL_EXPOSE_STATIC // Expose needed static functions for rccl-tests (or unit-testing in future)
 #ifdef RCCL_EXPOSE_STATIC
 #define rccl_static
 #define rccl_static_inline inline